meta-swi: Make TI Wilink device support optional
TI Wilink SDIO devices are supported through IoT extension slots of
MangOH boards. The drivers and supporting utilities for the same can be
kept in RootFS or be embedded in the application, according to use case.
A MACHINE_FEATURE, tiwifi, now inclues/excludes the support of TI Wilink
devices in RootFS.

- MACHINE_FEATURE tiwifi is added to mdm9x15 and mdm9x28 to maintain
  status quo.
- gpioexp utility which is MangOH board specific and not TI specific, is
now moved from sierra-init-tiwifi to i2cgpioctl recipe.
- i2cgpioctl itself is moved from common to meta-swi-mdm9xxx.
- Kernel configuration fragment support is enabled for WP85.
- tiwifi.cfg: Kernel configuration fragment to enable TI_WIFI in kernel.

Resolves: LXSWIREF-1239
Change-Id: Id5d4cd59808c96a3aeb8485e87b5f2bae04bdce0
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>