bluez5: separate append files from recipe files
This is necessary so that the files used by bluez_5.50.bb  aren't mixed
in with the files used by the bluez5_%.bbappend.  If the files are mixed
together, then a newer bluez recipe introduced by another layer can
accidentally pick up files associated with the 5.50 recipe because if
the newer recipe references a filename provided by the 5.50 recipe.
This happens because the bluez5_%.bbappend adds the path containing the
5.50 files to FILESEXTRAPATHS.

Change-Id: I35c8e1639fc134574a35d6abeb305cef3b72cb7c
Resolves: LXSWIREF-1641
Signed-off-by: David Frey <dfrey@sierrawireless.com>
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>