ecm: upgrade previously edited mdev.conf
When testing commit:
  cd01bd0 ecm: configure in response to hot plug event
it was discovered that systems may have already edited mdev.conf,
in which case the edited mdev.conf would overlay the changes from
above commit. As a result, ECM would miss its mdev.conf entry and
ECM interface would not show up.

Before overlay-mounting /etc, check if there is an overlaying RW
mdev.conf and if it contains ECM entries. If it does exist and
entries are absent, add the entries to the RW file. After that,
check if the RW and RO versions of files are identical and if yes,
delete the RW (top) version to allow the RO (bottom) version to
appear after overlay-mounting.

Resolves: LXSWIREF-1630
Related: LXSWIREF-1629
Change-Id: I49311c2a35470546f07a70f3da9127c66442e764
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>