security: include addition path of hdrcnv
Description:
It is an improvement ticket.
To align with the working environment in the signing server,
we have to add one more potential path of hdrcnv in swi-key-cwe.sh inside the server.
However, following the existing workflow, before updating swi-key-cwe.sh inside the server,
we have to update the below file in the firmware side:

meta-swi/meta-swi-mdm9xxx/recipes-bsp/android-signing/files/swi-key-cwe.sh

Root Cause:
In the signing server, if sign.sh has been executed,
it will trigger other script to execute the Linux executable hdrcnv.
However, hdrcnv may not be located correctly.
It may result in the failure of the script execution.

Solution:
Include the below lines in swi-key-cwe.sh:

if [ -x ../mdm9x28/common/sierra/hdrcnv ]; then
  HDRCNV="../mdm9x28/common/sierra/hdrcnv"
fi

Resolves: QTI9X07-4555
Change-Id: Ic23c10992c4df9c4182719b061dd46a38d3b9cc3
Signed-off-by: Lawrence Wong <lwo@sierrawireless.com>