network: smack: Ensure /etc/resolv.conf has SMACK label of "_"
In order to resolve the issue LE-13961, we need /etc/resolv.conf
to have a SMACK64 label of "_". It is best to have this file
created at or before the mount-early time during system boot.

For an IMA enable build, /etc/ is not a writeable dir. A soft link
to the /var/resolv.conf is then created to overcome this problem.
Current implementation, however, only creates the soft-link at
build time. We also need to create the source file at mount-early
time as well.

Resolves: LXSWIREF-1646
Related: LXSWIREF-1470 LE-13807 LE-13961
Change-Id: I6f3c5c39ec8bae5823c9325eb197735bf0c6a622