toolchain: Remove dependency on libcrypt-compat-dev
For 3rd party binaries (WEBSERVER in data), that require libcrypt.so.1,
libxcrypt-compat is added as RDEPENDS. Other packages that require
libcrypt RDEPENDS on "virtual/crypt" to get libcrypt.so.2. The way
libxcrypt-compat is implemented in Yocto 3.0 has a side effect in such
mixed configurations. In the SDK build, it results in duplication of
libcrypt.so and libxcrypt.so in dev packages of both libxcrypt and
libxcrypt-compat.

Considering that libxcrypt-compat-dev is not required and got included
as a side-effect, this package can be excluded from sdk build to avoid
the duplication issue.

Resolves: LXSWIREF-1649
Change-Id: Ia0a5f3b0cf47b2428a2696c6e4f1658c677a68c3
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>