common: workaround cve_check issue with Yocto 3.0
Patch https://patchwork.openembedded.org/patch/167782/ removed
"do_unpack" from the dependency of "do_cve_check". The assumption, "The
patch scanner works with patch files in the layer, not in the workdir,
so it doesn't need to unpack." does not seem correct. It appears to be
scanning all patch files listed in SRC_URI. In most cases the patch
files reside in the layer; but in case of recipes such as bash_5.0,
readline_5.2, there are patches to be downloaded. The downloaded patch
files reside in the workdir. When sstate is used and setscene tasks run
for do_populate_sysroot etc, unless cve_check declares dependency on
do_unpack, the patch files will not be downloaded.

This patch works-around the issue by adding back the dependency on the
impacted recipes. Once a fix is available in Yocto, this can be
reverted.

Resolves: LXSWIREF-1652
Change-Id: I87fb55ecf1b386a129e64e662b70fc608ada01c0
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>