wp76xx audio: Wait for audio stream to end.
It was observed that a very small tone at the end of a file
was not played. This was because as the audio path was closed
prior to the end of stream.
Adding a delay of 200 ms before PCM is closed resolved the
issue and the tone could be heard.

Resolves: QTI9X07-4837
Change-Id: Iede511eb365a3da0fa4dbeb7d241d9117d0bc2c6
Signed-off-by: infosys_ATMAINT <infosys_atmaint@sierrawireless.com>