smack: cipso: Disable netlabel if Legato is not started
When Legato isn't successfully started, "ping" command from target to host will
be "dropped". It is because the SMACK module in the target imposes CIPSO
settings which later be considered "parameter error" from the host.

Disabling netlabel can resolve this parameter error. When Legato fails to start
and SMACK is enabled in the kernel, netlabel and CIPSO will therefore be
explicitly disabled.

To prevent similar SMACK issue on a system build without Legato, setting the
SMACK64EXEC labels for system utils (dropbear, qmux, etc.) will be skipped in
the recipe.

Resolves: LXSWIREF-1682
Change-Id: Ia050d27098d928fec6a4fdafb21fd87e85c2eef9