security: validate inputs given to swi-key-cwe.sh
Description:
swi-key-cwe.sh does not validate the inputs that are given to it.
For example, Aptiv provided RFS*O* instead of RFS*0* and spent
time debugging why the RFS0 key was not installed properly.

Root Cause:
Script only checked if input arguments were not empty before
populating the parameters from arguments. Input validity checks
were not implemented in script code.

Solution:
Input parameter validation has been included in the script for
the 2nd parameter (productid - 9X28) and 3rd parameter (keyid
- [A-Z] for 1st 3 characters and [0-9] for the last character).
It has been implemented such that the code exits when there
is an invalid keyid. For productid, only 9X28 is supported
for now. As in the original script, the declaration of the
productid has been kept intact (i.e 9X28). However, a check has
been added while populating from arguments to show a warning
message in case of unsupported productid, without exiting from
the script and proceeding with generation of cwe with the
product value as 9X28.

Resolves: QTI9X07-4887
Change-Id: Ie68e895b7792fa761b3995da54853448de7cf46f
Signed-off-by: infosys_ATMAINT <infosys_atmaint@sierrawireless.com>