linux: bitbake 4.14 and wp76 recipe
Bitbake recipes for build the WP76 against the Linux 4.14 kernel.

Set the linux kernel build type for 4.x to linux-msm from the
from the kernel/Makefile. Don't try to get the version if there is
no kernel directory, for example virt.

Set default console kernel command line in kernel recipe and
allow it to be overriden in machine.conf file.

Limitations:
- linux 4.14 was missing msm_ipc.h

Resolves: LXSWIREF-1723
Change-Id: Ie0e59e9f3ffb2e17e72e192accd4eb2deca4821d
Signed-off-by: Steve MacDonald <smacdonald@sierrawireless.com>
----
[zmarkovic] Added staging of kernel headers. Fixed detection of
            kernel Makefile. Set default console device(s).