qcacld-hl: Correct 'Align to LE.UM.1.1-88300-9x07'
The commit 'Align to LE.UM.1.1-88300-9x07' is incorrect in the naming of
tag, as well as the implementation. The correct tag name is:
'LE.UM.3.4.2-01100-9x07'. Also the KERNEL_VERSION is not apt to select
SRC_REV. KERNEL_VERSION is set only after fetching the kernel sources;
it might not be set when SRC_REV of qcacld_hl is being evaluated. We
need to use a global env variable such as "PREFERRED_VERSION_linux-msm",
if present.

Upgrade qcacld-hl driver source to LE.UM.3.4.2-01100-9x07, so as to be
compatible with Linux kernel 4.14.

Keep the support for previous version until LE.UM.3.4.2-01100-9x07 is
fully validated in LXSWIREF-1722.

Resolves: LXSWIREF-1752
Relates: LXSWIREF-1722
Change-Id: I310a58064948964b8481f5af722cbceec34da554
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>