virt: generate 'scripts/mod/modpost' during toolchain install
The change https://patchwork.kernel.org/patch/10690901/ removed the
scripts/mod/ generation from 'make scripts' in the kernel sources.

As we need 'scripts/mod/modpost' for external module generation,
this adds back the modpost generation to 'make scripts', by dynamically
patching the kernel sources.

The point of making this patch dyanmic is to that this would hopefully
need less maintenance.

Resolves: LXSWIREF-1729
Change-Id: I52f81c2dd93694fecce4184cc8118965a081a9a5
Signed-off-by: Bertrand Roussel <broussel@sierrawireless.com>