yocto, audio: Fix QCOM's audio bb file
WORKSPACE variable points to apps_proc in image_src
build, but it is empty in image_bin build (customers
should not have the access to QCOM's code).

FILESPATH variable in audio bb file (taken as-is from
QCOM release 2.1.2) was improperly populated, causing
bitbake/quilt confusion when WORKSPACE variable is empty:

    FILESPATH = ": <other-path-parts>"

Unfortunately, there is no available Yocto method to
override this behavior in bbappend file, and original file
must be modified.

Resolves: LXSWIREF-1843
Change-Id: Ib572bcd49400b103b60e49b3bad9693b245624cd
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>