common: Implement xz compression options in populate_sdk
Bring in the fix from yocto-3.1. Description from original commit in
poky:

Building an SDK on a machine with 8GB RAM resulted in excessive swapping
due to the xz compressor using ~20GB of memory. This is because xz is
being called with "-T 0 -9".

To allow tuning the compression versus memory usage, introduce a variable
named SDK_XZ_OPTIONS that defaults to a more sane default:
SDK_XZ_OPTIONS ?= "${XZ_DEFAULTS} ${SDK_XZ_COMPRESSION_LEVEL}"
The use of XZ_DEFAULTS fixes the excessive memory usage.
The SDK_XZ_COMPRESSION_LEVEL variable allows overriding the speed vs
compression. In an office or development environment the extra time
spent on compressing a few percent more is just not worth it.

Resolves: LXSWIREF-1785
Related: LXSWIREF-1871
Change-Id: I094361bbc99a3b2f94d84cce156c8b1171e5eafc
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>