common: Reduce compression level for dev builds
The default compression level, set by poky, for use by xz in
do_populate_sdk, is "--memlimit=50% -9". 9 is the highest level of
compression, but is quite memory consuming (a bit more than 7GiB for a
2.2GiB SDK). --memlimit=50%, reduces this memory consumption by half but
doubles the time taken.

For release builds, it may be desirable to have high compression even
at the cost of increased time; however, for development builds such as
CI and local builds, a faster build with reduced compression levels
might be preferred.

This fix chooses "-T 0" option with xz, for all non-release builds. This
employs level 6 which is a good enough and faster compression. The
FW_VERSION variable is used to differentiate between a release and
non-release build, since this variable is set only during release
builds.

Resolves: LXSWIREF-1785
Change-Id: Ieae80818a7bc9b49a18a998d6a9dcbbd2bd5009d
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>