build.py: Introduce extra Yocto layers
Allow adding extra Yocto layers to the build by listing them in
EXTRA_LAYERS configuration variable. The syntax in this case is:
  EXTRA_LAYERS=<layer-list> make <target>

where <layer-list> is a colon (:) separated list of paths with
valid conf/layer.conf configuration files. If relative paths are
specified for layers, these will be calculated from top-level
directory.

Fix a bug in determining path to top directory which caused
optional meta-mangoh layer to be dropped from the build.

Resolves: LXSWIREF-1716
Change-Id: Iba26f7aeb751066e7a4f0174a769ae3e42a16b05
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>