meta-swi-mdm9xxx: Support high-speed on UART2
Update functions.env to lookup for /dev/ttyHS1, which is high-speed
services mapping on UART2.

Check for "UNAVAILABLE" config is removed, since the block is never
entered. When a configuration is "UNAVAILABLE", the driver probe fails
and there will not be any /sys/class/tty entry made for the UART.

Resolves: LXSWIREF-1894
Change-Id: I31d9f5b01a25fabacf93cc3439a8cc7b293b39cd
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>