mdm9x28: Remove unused restart_at_uart script
On WP modules, when UART is configured for AT service, management of
the interface is done by the modem. Only in AR modules, the application
processor manages the interface; and only there the AT Service
initialization on UART is required to be done in Linux.

This commit limits restart_at_uart to AR and removes it from the common
build.

Resolves: LXSWIREF-1894
Change-Id: I037caf48433d656207b0d84036bbbc8b1ab4eb6a
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>