fx30: Add dependency on WP team's layers
Add the WP product team's layer as dependency for fx30 product layer in
meta-swi and meta-swi-extras. This ensures:
 - WP layers, if present are pulled in by build.py during fx30 build.
 - WP layers are placed below fx30 layers in BBLAYERS.
Multiple dependencies from a single layer are added as space separated
entries with lower priority layer first. This ensures that, the higher
priority layer is placed above the lower priority layer in BBLAYERS.
  meta-swi-extras/meta-swi-mdm9x28-fx30
  meta-swi-wp-extras/meta-swi-wp
  meta-swi-extras/meta-swi-mdm9x28-bin
  :
  meta-swi/meta-swi-mdm9x28-fx30
  meta-swi-wp/meta-swi-wp
  meta-swi/meta-swi-mdm9x28-wp
  meta-swi/meta-swi-mdm9x28

Resolves: LXSWIREF-1979
Change-Id: I92161b2f5c9c7a21ac656a59e6ff305cb8bc6f93
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>