dbus: Restore path of system_bus_socket to /var/run/dbus/
Commit 927e154 in Yocto-3.1, poky directory changed the path of
system_bus_socket from /var/run/dbus/ to /run/dbus/. This is based on
the assumption that either /var/run does not exist or if it does, it
is same as /run, which is the case for most of the latest systems. This
is explained in https://gitlab.freedesktop.org/dbus/dbus/-/issues/282.

However on our targets, including virt, this is not the case. /var/run
and /run both exist, and they are different. Hence, this fix needs to be
reverted using an append in meta-swi.

Resolves: LXSWIREF-1926
Change-Id: I75a8fbcb1eeaae7bf576f3e241dae49aabc7aaa3
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>