bluez5: Install a default /etc/bluetooth/main.conf
Recent BlueZ enhancements are changing the default behaviour of
bluetoothd with options to tune those behaviours through
/etc/bluetooth/main.conf.

For example, the change def01de025 introduced "TemporaryTimeout" that
affects how long a device information is kept available during or after
scanning. Earlier, the device information was available for up to 180
seconds (3 minutes) after scanning is turned off. The new behaviour is
that, the removal will be after "TemporaryTimeout" seconds of not seeing
the device, even when scanning is ON. The default being 30 seconds.

Since, more such configurations are being introduced, it will be
easier for user if a default /etc/bluetooth/main.conf exists which they
can modify to suit their requirements. Update the recipe append to
install the default main.conf from the bluez sources into
/etc/bluetooth.

Resolves: LXSWIREF-2002
Change-Id: Idb879de50e347ab6d22ca391d9a096f897552073
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>