sdk: Make SDKPATH same as SDKPATHINSTALL
Decoupling of SDK build path(SDKPATH), and the SDK installation
path(SDKPATHINSTALL), requires the relocation framework such as
environment script in place, which are only available the SDK shar(.sh),
which is the nominal output of Yocto. SWI distro also ouput a tar.xz
which contains only the sysroots and this does not support relocation.

Due to the absence of relocation framework in tar.xz, the paths to SDK
embedded into the binaries and executables at build time will not be
updated to the installation path and hence the executables will fail to
run with broken paths, if the SDK is installed at a different path.

This means that the user of tar.xz needs to install it at the $SDKPATH.
Update it to the familiar /opt/swi location, which is the default
installation path ($SDKPATHINSTALL).

Resolves: LXSWIREF-2040
Change-Id: Ie0f9c5e1f72cfb2eb66d148bae9cac047d5ad361
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>