fakeroot: Enable build with yocto-4.0
When fakeroot tool is built for yocto-4.0 SDK, libfakeroot.so is
linked against yocto's libc library. However, when SDK's fakeroot
is executed, libfakeroot.so attempts to load host's libc library
which does not support all symbols defined in yocto's libc. This
causes run-time error.

Use build host's compiler to generate fakeroot and link to host's
libc. This generates code that is backwards compatible to host's
linux release.

Use default sysv_ipc to communicate to faked. This removes the
need for networking and works around the failed loopback connect
request. This is a temporary solution until a better solution is
found.

Resolves: LXSWIREF-2071
Change-Id: I0ee855832dd1e9a95e519b589f192228b79dec6e
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>