initscripts: Fix usage of basename from current shell
Environment variable $0 representing the current shell is prefixed
with hyphen (-) if the current shell is a login shell. When this
is used in 'basename $0', basename confuses expansion of $0 to -sh
with a command line option -s SUFFIX, as shown by warning message:
  Usage: basename FILE [SUFFIX] | -a FILE... | -s SUFFIX FILE...

Prefix $0 with two hyphens (--) to enforce interpretation of $0
as current shell in expansion of 'basename -- $0'.

Resolves: LXSWIREF-2071
Change-Id: I8256307775ae4a83d6acf50f90213a612c6281ec
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>