swi-version: Use manifest link instead of file
Builds parse the manifest file to determine linux kernel version.
Manifest files use the DATETIME environment variable to append to
file name, which makes their name dependent on when the build task
started, as shown in this example:
 $ bitbake -e mdm9x28-image-minimal |grep "^IMAGE_VERSION_SUFFIX="
 IMAGE_VERSION_SUFFIX="-20220601224908"
 $ bitbake -e mdm9x28-image-minimal |grep "^IMAGE_VERSION_SUFFIX="
 IMAGE_VERSION_SUFFIX="-20220601224917"

This is a problem because the manifest file name determined in one
build step may not match the manifest file generated in another
build step. Fortunately, the build also creates a time-independent
symbolic link to the manifest file.

Parse the symbolic link instead of actual manifest file to remove
dependancy on the value of DATETIME environment variable.

Resolves: FARM-6425
Change-Id: Ia68897c226338c15cc736f80fa6fc94c651c3865