fakeroot: Build with yocto-4.0 compiler and libraries
Fakeroot used build host's compiler for building/installing it to
yocto SDK. While this worked if the SDK is installed into the same
system as the build host, this is still incorrect since SDK target
host may be a different system type. For example, an x86_64 host
could be generating SDKs for both x86_64 and i686 hosts.

Additional problem was that all libdl functions were moved to libc
sometime around version 2.33, and glibc package stopped creating a
symlink libdl.so -> libdl.so.2. This caused all packages to link
libdl.a statically, eliminating the need for linking it at run
time. When this is run on systems with glibc version prior to
2.33, it failed due to missing symbols dlopen and dlsym.

Use yocto's cross-compiler to generate fakeroot binary for SDK.
Depending on SDK target, this could be a x86_64 or i686 compiler.
Upgrade fakeroot to version 1.29 for linking with recent versions
of glibc. Create link libdl.so -> libdl.so.2 in recipe-sysroot to
force dynamic linking of the library, both at build time and run
time. Avoid hard-coding libdl symbol versions to enable linking
to older versions of libc and libdl at run time.

Resolves: LXSWIREF-2077
Change-Id: Id7be2eb63b3e68a68e3634499c0b3b33307b7d30
Relates: FARM-6425
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>