mdm9x28: Remove ubifs image creation when QCA is enabled
The rootfs image size has grown to a point where the ubifs image
exceeds the system partition size when QCA is enabled. This patch
removes ubifs image creation when QCA is enabled, as a work-around to
let the build pass. The restriction imposed is that, only squashfs image
is available to a user of QCA enabled build. A QCA disabled build will
continue to produce both squashfs and ubifs images.

Resolves: LXSWIREF-2078
Related: LXSWIREF-2082
Change-Id: I7659232685839096bd2bc1e3f919ed3ab74e3957
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>